+++ api/tasks/celery.py
@@ -3,7 +3,7 @@ from celery import Celery
 from api.config import Config
 
 celery = Celery(
-    'celery',
+    'tasks',
     broker=f'redis://{Config.REDIS_HOST}:{Config.REDIS_PORT}',
-    include=['app.celery_tasks']
+    include=['api.tasks.tasks']
 )
